Refactor ResearchPage component by removing hardcoded research data and filtering logic; integrate ResearchClient for enhanced interactivity and maintainability.
